third_party/WebKit/LayoutTests/performance/event-timing/event-timing-bufferbeforeonload.html

<!DOCTYPE html>
<html>
<meta charset=utf-8 />
<title>Event Timing: buffer long-latency events before onload</title>
<button id='button' onclick='1'>Generate a 'click' event</button>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<script src=./resources/event-timing-support.js></script>

<script>
  let clickTimeMin;
  let processingStartMin;

  function validateEntries() {
    const entries = performance.getEntriesByName('click');
    const onloadTime = performance.timing.loadEventStart - performance.timeOrigin;

    const entriesBeforeOnload = entries.filter(e => e.startTime <
      onloadTime);
    assert_equals(entriesBeforeOnload.length, 1,
      "Long latency events before onload should be buffered."
    );
    const entry = entriesBeforeOnload[0];
    verifyClickEvent(entry);

    assert_greater_than(entry.processingStart, processingStartMin);
    assert_greater_than(onloadTime, entry.processingStart);
    assert_greater_than(entry.startTime, clickTimeMin);
    assert_greater_than(onloadTime, entry.startTime);

    const entriesAfterOnload = entries.filter(e => e.startTime >
      onloadTime);
    assert_equals(entriesAfterOnload.length, 0,
      "Events after onload shouldn't be buffered.");
  }

  /* Timeline:
     Begin Busy Loop
     Click 1 arrives
     End Busy Loop
     (Dispatch and Process Click 1 - buffered)
     Onload Event Fires
     Begin Busy Loop
     Click 2 arrives
     End Busy Loop
     (Dispatch and Process Click 2 - not buffered)
  */
  async_test(function(t) {
    clickTimeMin = performance.now();
    clickAndBlockMain('button');
    processingStartMin = performance.now();
    on_event(window, 'load', e => {
      clickAndBlockMain('button').then(wait).then(
          t.step_func_done(validateEntries));
    });
  }, "Event Timing: click, onload.");

</script>
</html>
